Return JSON attachments to the caller as a byte stream.
The http module no longer automatically decodes 'application/json' responses,
instead leaving it to the caller, i.e. the client module, to decide (because
only the caller really knows what the response means).

To support the client module, the http.Resource class now includes a set of
methods (get_json, post_json, etc) that explicitly read and decode the
response.  The methods also ensures that the response is fully read.